use SchemaSourceDao in the FilesResource.java
